Relax grammar pins and upgrade tree-sitter now that rca is gone
ratchet defines its own metrics, so the grammars no longer need to match the
exact versions rust-code-analysis used. Relax the `=0.23.x` pins to caret ranges
and bump to the latest compatible releases:

- tree-sitter core 0.25 -> 0.26, tree-sitter-rust 0.23.2 -> 0.23.3,
  tree-sitter-python 0.23.6 -> 0.25.0 (cpp/java/typescript already latest).
- tree-sitter 0.26 changed `Node::child`/`named_child` to take `u32`; the loop
  counters stay `usize` for the still-`usize` `*_count` methods and cast at the
  child access (`i as u32`).

No metric shift: the golden fixture test and the committed quality-report.json
are both unchanged across the upgrade, so no re-bless was needed. fmt/clippy/test
all clean.

+++ src/native/analysis.rs
@@ -51,7 +51,7 @@ pub fn visit_functions(rules: &Rules, tree: &Tree, source: &[u8], f: &mut impl F
     fn recurse(walk: &mut Walk, node: Node, f: &mut impl FnMut(&str, Node)) {
         let mut i = 0;
         while i < node.named_child_count() {
-            let child = node.named_child(i).expect("named_child within count");
+            let child = node.named_child(i as u32).expect("named_child within count");
             if walk.rules.is_function(child.kind()) {
                 let name = name_for(walk.rules, &child, walk.source, &mut walk.unnamed);
                 f(&name, child);
@@ -78,7 +78,7 @@ pub fn file_functions(rules: &Rules, tree: &Tree, _source: &[u8]) -> u64 {
     fn recurse(rules: &Rules, node: Node, count: &mut u64) {
         let mut i = 0;
         while i < node.named_child_count() {
-            let child = node.named_child(i).expect("named_child within count");
+            let child = node.named_child(i as u32).expect("named_child within count");
             if rules.counts_toward_nom(child.kind()) {
                 *count += 1;
             }
@@ -140,7 +140,7 @@ fn own_params(node: &Node, rules: &Rules) -> u64 {
 fn sum_closure_params(node: &Node, rules: &Rules, total: &mut u64) {
     let mut i = 0;
     while i < node.named_child_count() {
-        let child = node.named_child(i).expect("named_child within count");
+        let child = node.named_child(i as u32).expect("named_child within count");
         let kind = child.kind();
         if rules.is_function(kind) {
             // A nested space accounts for its own closures.
